+++ src/middlewares/mappers/ignore-expiration.js
@@ -1,4 +1,0 @@
-export function ignoreExpiration(_req, res, next) {
-  res.locals.tokens = { ...res.locals.tokens, ignoreExpiration: true };
-  next();
-}